A few major changes
Has to revert to the old movement system for the player at this point because the tornado direction wouldnt get properly called.

Adds more notes.

in the Tornado.gd there are 3 seperate entire files of how to do do the tutorial. the bottom commented ones arent functioning properly, so they are left commented. but you can change the vars to how it includes Globals.gd